requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-gitlab>=3.15.0
claude-agent-sdk>=0.1.0
pyyaml>=6.0
python-dotenv>=1.0.0
